chore: refactoring MapList toolbars
